Collect trivial conversion fixes
These are very local, simple fixes that do not
warrant their own commit.

Acked-by: Gert Doering <gert@greenie.muc.de>
Message-Id: <20250817143922.17988-1-gert@greenie.muc.de>
URL: https://www.mail-archive.com/openvpn-devel@lists.sourceforge.net/msg32611.html

+++ src/openvpn/init.c
@@ -343,7 +343,7 @@ management_callback_remote_entry_get(void *arg, unsigned int index, char **remot
         const char *status = (ce->flags & CE_DISABLED) ? "disabled" : "enabled";
 
         /* space for output including 3 commas and a nul */
-        int len =
+        size_t len =
             strlen(ce->remote) + strlen(ce->remote_port) + strlen(proto) + strlen(status) + 3 + 1;
         char *out = malloc(len);
         check_malloc_return(out);
